--- Assets/LeapMotion/Core/Scripts/Utils/BitConverterNoAlloc.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 9f0f717847604974ebed898e5ba31240
-timeCreated: 1510945017
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
