+++ Assets/LeapMotion/Core/Scripts/Utils/BitConverterNonAlloc.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
 guid: 05cd563a8e42858468aa2d4f87b48e7e
-timeCreated: 1503513952
+timeCreated: 1510945786
 licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
