+++ Assets/LeapMotion/Generation/BitConverter/BitConverterTemplate.cs
@@ -2,7 +2,7 @@
 
 namespace Leap.Unity.Generation {
 
-  public static class BitConverterNoAlloc {
+  public static class BitConverterNonAlloc {
     //BEGIN
 
     public static _Primitive_ To_Primitive_(byte[] bytes, int offset = 0) {
@@ -24,7 +24,7 @@ namespace Leap.Unity.Generation {
       }
     }
 
-    public static void ToBytes(_Primitive_ primitive, byte[] bytes, int offset = 0) {
+    public static void GetBytes(_Primitive_ primitive, byte[] bytes, int offset = 0) {
       unsafe {
         fixed (byte* ptr = &bytes[offset]) {
           _Primitive_* primitivePtr = (_Primitive_*)ptr;
@@ -33,7 +33,7 @@ namespace Leap.Unity.Generation {
       }
     }
 
-    public static void ToBytes(_Primitive_ primitive, byte[] bytes, ref int offset) {
+    public static void GetBytes(_Primitive_ primitive, byte[] bytes, ref int offset) {
       unsafe {
         fixed (byte* ptr = &bytes[offset]) {
           offset += sizeof(_Primitive_);
